change to prism as feature extractor for pathology
